# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  featuredImage?: string | null;
  techStack: string[];
  category: string;
  year: number;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  teamSize?: number | null;
  featured: boolean;
  /** @nullable */
  repositoryLink?: string | null;
  /** @nullable */
  liveLink?: string | null;
  /** @nullable */
  caseStudySlug?: string | null;
  createdAt?: string;
}

export interface ProjectDetail {
  id: number;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  featuredImage?: string | null;
  galleryImages?: string[];
  techStack: string[];
  category: string;
  year: number;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  teamSize?: number | null;
  featured: boolean;
  /** @nullable */
  repositoryLink?: string | null;
  /** @nullable */
  liveLink?: string | null;
  /** @nullable */
  caseStudySlug?: string | null;
  createdAt?: string;
}

export interface StatCount {
  label: string;
  count: number;
}

export interface ProjectStats {
  totalCount: number;
  byCategory: StatCount[];
  byYear: StatCount[];
  byTech: StatCount[];
}

export interface CaseStudy {
  id: number;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectSlug?: string | null;
  /** @nullable */
  projectTitle?: string | null;
  slug: string;
  title: string;
  /** @nullable */
  challenge?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  technicalDetails?: string | null;
  /** @nullable */
  results?: string | null;
  /** @nullable */
  learnings?: string | null;
  /** @nullable */
  processImage?: string | null;
  /** @nullable */
  architectureDiagram?: string | null;
  /** @nullable */
  resultsImage?: string | null;
  createdAt?: string;
}

export interface GalleryImage {
  id: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  category: string;
  tags?: string[];
  /** @nullable */
  altText?: string | null;
  createdAt?: string;
}

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  author?: string | null;
  category: string;
  tags?: string[];
  /** @nullable */
  featuredImage?: string | null;
  /** @nullable */
  readingTimeMinutes?: number | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
  budget?: string;
  timeline?: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  message: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  timeline?: string | null;
  readStatus: boolean;
  /** @nullable */
  respondedAt?: string | null;
  createdAt: string;
}

export interface SubscribeInput {
  email: string;
  name?: string;
}

export interface SubscribeResult {
  success: boolean;
  message: string;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  level: string;
  /** @nullable */
  yearsOfExperience?: number | null;
}

export interface SkillGroup {
  category: string;
  skills: Skill[];
}

export type ListProjectsParams = {
category?: string;
year?: number;
tech?: string;
featured?: boolean;
};

export type ListGalleryParams = {
category?: string;
};

export type ListBlogPostsParams = {
category?: string;
tag?: string;
search?: string;
};