# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectDetail {
  id: number;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  featuredImage?: string | null;
  galleryImages?: string[];
  techStack: string[];
  category: string;
  year: number;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  teamSize?: number | null;
  featured: boolean;
  /** @nullable */
  repositoryLink?: string | null;
  /** @nullable */
  liveLink?: string | null;
  /** @nullable */
  caseStudySlug?: string | null;
  createdAt?: string;
}