# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API
 * OpenAPI spec version: 0.1.0
 */
import type { StatCount } from './statCount';

export interface ProjectStats {
  totalCount: number;
  byCategory: StatCount[];
  byYear: StatCount[];
  byTech: StatCount[];
}