# lib/api-zod/src/generated/types/subscribeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeInput {
  email: string;
  name?: string;
}