# lib/api-zod/src/generated/types/subscribeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeResult {
  success: boolean;
  message: string;
}